requirements.txt
transformers
nonebot-plugin-htmlrender
nonebot-adapter-onebot
beautifulsoup4
pynvml